+++ src/worker/worker.runtime.test.ts
@@ -128,6 +128,7 @@ class FakeWorkerGateway {
   readonly acceptedTranscriptRequests: WorkerTranscriptCommitParams[] = [];
   readonly liveEventRequests: WorkerLiveEventParams[] = [];
   readonly inferenceRequests: WorkerInferenceStartParams[] = [];
+  readonly applicationOrder: string[] = [];
 
   constructor(private readonly options: FakeGatewayOptions = {}) {
     this.httpServer = createServer();
@@ -315,6 +316,7 @@ class FakeWorkerGateway {
       return;
     }
     this.acceptedTranscriptRequests.push(structuredClone(frame.params));
+    this.applicationOrder.push(`transcript:${frame.params.seq}`);
     this.send(socket, {
       type: "res",
       id: frame.id,
@@ -331,6 +333,11 @@ class FakeWorkerGateway {
   private handleLiveEvent(socket: WebSocket, frame: WorkerLiveEventRequestFrame): void {
     this.methods.push(frame.method);
     this.liveEventRequests.push(structuredClone(frame.params));
+    this.applicationOrder.push(
+      frame.params.event.kind === "lifecycle"
+        ? `live:lifecycle:${frame.params.event.payload.phase}`
+        : `live:${frame.params.event.kind}`,
+    );
     if (this.options.silenceFirstLiveEvent && !this.droppedLiveEvent) {
       this.droppedLiveEvent = true;
       return;
@@ -706,6 +713,7 @@ function descriptor(socketPath: string, workspaceDir: string): WorkerLaunchDescr
       runId: RUN_ID,
       turnId: "worker-turn",
       prompt: "Complete the worker turn.",
+      suppressPromptTranscript: false,
       workspaceDir,
       modelRef: MODEL_REF,
       inferenceOptions: { reasoning: "off" },
@@ -754,6 +762,14 @@ describe("worker runtime", () => {
     expect(gateway.inferenceRequests[0]?.context.systemPrompt).toContain("worker-bootstrap-marker");
     const toolNames = gateway.inferenceRequests[0]?.context.tools?.map((tool) => tool.name) ?? [];
     expect(toolNames).toHaveLength(6);
+    const terminalIndex = gateway.applicationOrder.findIndex(
+      (entry) => entry === "live:lifecycle:end",
+    );
+    const finalTranscriptIndex = gateway.applicationOrder.findLastIndex((entry) =>
+      entry.startsWith("transcript:"),
+    );
+    expect(finalTranscriptIndex).toBeGreaterThanOrEqual(0);
+    expect(terminalIndex).toBeGreaterThan(finalTranscriptIndex);
     expect(toolNames).toEqual(
       expect.arrayContaining(["read", "write", "edit", "apply_patch", "exec", "process"]),
     );
@@ -806,7 +822,7 @@ describe("worker runtime", () => {
       gateway.liveEventRequests.some(
         (request) => request.event.kind === "lifecycle" && request.event.payload.phase === "error",
       ),
-    ).toBe(true);
+    ).toBe(false);
   });
 
   it("renumbers live events after a gateway cursor reset without aborting the run", async () => {
@@ -824,10 +840,10 @@ describe("worker runtime", () => {
     expect(gateway.liveEventRequests[1]?.event).toEqual(gateway.liveEventRequests[0]?.event);
   });
 
-  it("degrades hard live-event failures without affecting inference or transcript commits", async () => {
+  it("requires authoritative terminal delivery after degrading preview live events", async () => {
     const { gateway, launch } = await setup({ liveFailure: "capacity-exceeded" });
 
-    await expect(runWorkerDescriptor(launch)).resolves.toMatchObject({ status: "completed" });
+    await expect(runWorkerDescriptor(launch)).rejects.toThrow("worker live event rejected");
 
     expect(gateway.inferenceRequests).toHaveLength(1);
     expect(
@@ -835,7 +851,11 @@ describe("worker runtime", () => {
         .flatMap((request) => request.messages)
         .map((message) => message.role),
     ).toEqual(["user", "assistant"]);
-    expect(gateway.liveEventRequests).toHaveLength(1);
+    expect(gateway.liveEventRequests.length).toBeGreaterThanOrEqual(2);
+    expect(gateway.liveEventRequests.at(-1)?.event).toMatchObject({
+      kind: "lifecycle",
+      payload: { phase: "end" },
+    });
   });
 
   it("degrades a repeated no-progress live resync without hanging the run", async () => {
@@ -849,7 +869,11 @@ describe("worker runtime", () => {
 
     expect(gateway.inferenceRequests).toHaveLength(1);
     expect(gateway.acceptedTranscriptRequests).toHaveLength(2);
-    expect(gateway.liveEventRequests).toHaveLength(2);
+    expect(gateway.liveEventRequests).toHaveLength(3);
+    expect(gateway.liveEventRequests.at(-1)?.event).toMatchObject({
+      kind: "lifecycle",
+      payload: { phase: "end" },
+    });
   });
 
   it("fails closed when worker admission is rejected", async () => {
@@ -887,6 +911,10 @@ describe("worker runtime", () => {
 
     await expect(result).rejects.toThrow("operator stopped worker");
     expect(gateway.methods).toContain("worker.inference.cancel");
+    expect(gateway.liveEventRequests.at(-1)?.event).toMatchObject({
+      kind: "lifecycle",
+      payload: { phase: "end", aborted: true },
+    });
   });
 
   it("bounds shutdown when remote inference cancellation cannot settle", async () => {
@@ -905,14 +933,17 @@ describe("worker runtime", () => {
   });
 
   it.each([
-    ["error", "fixture provider failed", "error", "error"],
-    ["cancelled", "fixture inference cancelled", "aborted", "end"],
+    ["error", "error", "error"],
+    ["cancelled", "aborted", "end"],
   ] as const)(
     "reports remote inference %s terminals as failed turns",
-    async (plan, message, stopReason, lifecyclePhase) => {
+    async (plan, stopReason, lifecyclePhase) => {
       const { gateway, launch } = await setup({ inferencePlans: [plan] });
 
-      await expect(runWorkerDescriptor(launch)).rejects.toThrow(message);
+      await expect(runWorkerDescriptor(launch)).resolves.toEqual({
+        status: "failed",
+        reason: "turn-failed",
+      });
       const assistant = gateway.transcriptRequests
         .flatMap((request) => request.messages)
         .toReversed()
@@ -925,6 +956,19 @@ describe("worker runtime", () => {
       expect(lifecycle).toMatchObject({ payload: { phase: lifecyclePhase } });
     },
   );
+
+  it("keeps an unacknowledged failed-turn terminal as an infrastructure failure", async () => {
+    const { gateway, launch } = await setup({
+      inferencePlans: ["error"],
+      liveFailure: "capacity-exceeded",
+    });
+
+    await expect(runWorkerDescriptor(launch)).rejects.toThrow("worker live event rejected");
+    expect(gateway.liveEventRequests.at(-1)?.event).toMatchObject({
+      kind: "lifecycle",
+      payload: { phase: "error" },
+    });
+  });
 
   it("fails closed when a heartbeat is rejected without fencing", async () => {
     const { launch } = await setup({
@@ -958,9 +1002,10 @@ describe("worker runtime", () => {
     async (plan) => {
       const { gateway, launch } = await setup({ inferencePlans: [plan] });
 
-      await expect(runWorkerDescriptor(launch)).rejects.toThrow(
-        "Worker inference result exceeds the transcript message limit.",
-      );
+      await expect(runWorkerDescriptor(launch)).resolves.toEqual({
+        status: "failed",
+        reason: "turn-failed",
+      });
       const assistant = gateway.transcriptRequests
         .flatMap((request) => request.messages)
         .toReversed()
